requirements.txt
django==4.0.5
djangorestframework==3.13.1
pytest>=7.1.2
pytest-django>=4.5.2
pytest-cov>=3.0.0